pyproject.toml
[project]
name = "ebay-tracker"
version = "0.1.0"
description = "A script that tracks eBay item prices over time"
readme = "README.md"
requires-python = ">=3.10"
[tool.pytest.ini_options]
testpaths = ["tests"]
[dependency-groups]
dev = ["pytest >= 8.0", "pytest-mock >= 3.0"]
dependencies = [
    "beautifulsoup4==4.12.2",
    "numpy==1.26.2",
    "playwright==1.49.1",
    "tenacity>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]
